+++ pybamm/input/parameters/lithium-ion/cathodes/NMC_UMBL_Mohtat2020/NMC_electrolyte_exchange_current_density_PeymanMPM.py
@@ -1,0 +1,35 @@
+from pybamm import exp, constants, standard_parameters_lithium_ion
+
+
+def NMC_electrolyte_exchange_current_density_PeymanMPM(c_e, c_s_surf, T):
+    """
+    Exchange-current density for Butler-Volmer reactions between NMC and LiPF6 in
+    EC:DMC.
+
+    References
+    ----------
+    .. Peyman MPM manuscript (to be submitted)
+
+    Parameters
+    ----------
+    c_e : :class:`pybamm.Symbol`
+        Electrolyte concentration [mol.m-3]
+    c_s_surf : :class:`pybamm.Symbol`
+        Particle concentration [mol.m-3]
+    T : :class:`pybamm.Symbol`
+        Temperature [K]
+
+    Returns
+    -------
+    :class:`pybamm.Symbol`
+        Exchange-current density [A.m-2]
+    """
+    m_ref = 4.824 * 10 ** (-6)  # (A/m2)(mol/m3)**1.5 - includes ref concentrations
+    E_r = 39570
+    arrhenius = exp(E_r / constants.R * (1 / 298.15 - 1 / T))
+
+    c_p_max = standard_parameters_lithium_ion.c_p_max
+
+    return (
+        m_ref * arrhenius * c_e ** 0.5 * c_s_surf ** 0.5 * (c_p_max - c_s_surf) ** 0.5
+    )