+++ pybamm/parameters/parameter_sets.py
@@ -18,6 +18,7 @@ NCA_Kim2011 = {
     "cathode": "nca_Kim2011",
     "electrolyte": "lipf6_Kim2011",
     "experiment": "1C_discharge_from_full_Kim2011",
+    "sei": "Example",
     "citation": "kim2011multi",
 }
 
@@ -40,6 +41,7 @@ Marquis2019 = {
     "cathode": "lico2_Marquis2019",
     "electrolyte": "lipf6_Marquis2019",
     "experiment": "1C_discharge_from_full_Marquis2019",
+    "sei": "Example",
     "citation": "marquis2019asymptotic",
 }